requirements.txt
lxml
numpy
pandas
matplotlib
torch
torchvision
Pillow
opencv-python
resnest
efficientnet_pytorch